Don't throw away non-hideable items before cache
For example, the AudioVideoModule is now cached where it wasn't. When it appeared in the menu bar, it would ruin the cached images, as the crop frame of each image was off by the width of the item. Now, both the item and its image are cached, but we choose not to display it