Fix a memory issue
The update flags were not being cleared. Since all flags are added when the panel is shown, every time the `performUpdates` method ran, it would perform all updates, regardless of whether it was supposed to. This was causing brief, but substantial memory spikes.

This commit also removes unnecessary `async` and `throws` annotations that were missed in a previous refactoring pass, and improves the efficiency of the menu bar average color cache.